cherry-pick\:efb6aac17… :wrench: Use current_app in airtable_client.py
